fix: fix compilation and licenses
