--- operate/schema/src/test/java/io/camunda/operate/store/ImportStoreIT.java
@@ -1,98 +1,0 @@
-/*
- * Copyright Camunda Services GmbH
- *
- * BY INSTALLING, DOWNLOADING, ACCESSING, USING, OR DISTRIBUTING THE SOFTWARE (“USE”), YOU INDICATE YOUR ACCEPTANCE TO AND ARE ENTERING INTO A CONTRACT WITH, THE LICENSOR ON THE TERMS SET OUT IN THIS AGREEMENT. IF YOU DO NOT AGREE TO THESE TERMS, YOU MUST NOT USE THE SOFTWARE. IF YOU ARE RECEIVING THE SOFTWARE ON BEHALF OF A LEGAL ENTITY, YOU REPRESENT AND WARRANT THAT YOU HAVE THE ACTUAL AUTHORITY TO AGREE TO THE TERMS AND CONDITIONS OF THIS AGREEMENT ON BEHALF OF SUCH ENTITY.
- * “Licensee” means you, an individual, or the entity on whose behalf you receive the Software.
- *
- * Permission is hereby granted, free of charge, to the Licensee obtaining a copy of this Software and associated documentation files to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject in each case to the following conditions:
- * Condition 1: If the Licensee distributes the Software or any derivative works of the Software, the Licensee must attach this Agreement.
- * Condition 2: Without limiting other conditions in this Agreement, the grant of rights is solely for non-production use as defined below.
- * "Non-production use" means any use of the Software that is not directly related to creating products, services, or systems that generate revenue or other direct or indirect economic benefits.  Examples of permitted non-production use include personal use, educational use, research, and development. Examples of prohibited production use include, without limitation, use for commercial, for-profit, or publicly accessible systems or use for commercial or revenue-generating purposes.
- *
- * If the Licensee is in breach of the Conditions, this Agreement, including the rights granted under it, will automatically terminate with immediate effect.
- *
- * SUBJECT AS SET OUT BELOW, THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- * NOTHING IN THIS AGREEMENT EXCLUDES OR RESTRICTS A PARTY’S LIABILITY FOR (A) DEATH OR PERSONAL INJURY CAUSED BY THAT PARTY’S NEGLIGENCE, (B) FRAUD, OR (C) ANY OTHER LIABILITY TO THE EXTENT THAT IT CANNOT BE LAWFULLY EXCLUDED OR RESTRICTED.
- */
-package io.camunda.operate.store;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import io.camunda.operate.JacksonConfig;
-import io.camunda.operate.Metrics;
-import io.camunda.operate.conditions.DatabaseInfo;
-import io.camunda.operate.connect.ElasticsearchConnector;
-import io.camunda.operate.connect.OperateDateTimeFormatter;
-import io.camunda.operate.property.OperateProperties;
-import io.camunda.operate.schema.AbstractSchemaIT;
-import io.camunda.operate.schema.SchemaManager;
-import io.camunda.operate.schema.elasticsearch.ElasticsearchSchemaManager;
-import io.camunda.operate.schema.indices.ImportPositionIndex;
-import io.camunda.operate.schema.util.SchemaTestHelper;
-import io.camunda.operate.schema.util.TestTemplate;
-import io.camunda.operate.schema.util.elasticsearch.ElasticsearchSchemaTestHelper;
-import io.camunda.operate.schema.util.opensearch.OpenSearchSchemaTestHelper;
-import io.camunda.operate.store.elasticsearch.ElasticsearchImportStore;
-import io.camunda.operate.store.elasticsearch.ElasticsearchTaskStore;
-import io.camunda.operate.store.elasticsearch.RetryElasticsearchClient;
-import io.micrometer.core.instrument.MeterRegistry;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.ContextConfiguration;
-
-@ContextConfiguration(
-    classes = {
-      ImportPositionIndex.class,
-      TestTemplate.class,
-      ElasticsearchSchemaManager.class,
-      ElasticsearchImportStore.class,
-      OpenSearchSchemaTestHelper.class,
-      ElasticsearchSchemaTestHelper.class,
-      DatabaseInfo.class,
-      OperateProperties.class,
-      RetryElasticsearchClient.class,
-      ElasticsearchConnector.class,
-      ElasticsearchTaskStore.class,
-      JacksonConfig.class,
-      OperateDateTimeFormatter.class,
-      Metrics.class
-    })
-public class ImportStoreIT extends AbstractSchemaIT {
-
-  @Autowired public ImportStore importStore;
-
-  @Autowired public SchemaManager schemaManager;
-
-  @Autowired public ImportPositionIndex importPositionIndex;
-
-  @Autowired public SchemaTestHelper schemaHelper;
-
-  @MockBean public MeterRegistry meterRegistry;
-
-  @BeforeEach
-  public void createDefault() {
-    schemaManager.createDefaults();
-  }
-
-  @AfterEach
-  public void dropSchema() {
-    schemaHelper.dropSchema();
-  }
-
-  @Test
-  public void setAndGetConcurrencyMode() {
-    schemaManager.createIndex(
-        importPositionIndex, "/schema/elasticsearch/create/index/operate-import-position.json");
-
-    assertThat(importStore.getConcurrencyMode()).isFalse();
-
-    importStore.setConcurrencyMode(true);
-    assertThat(importStore.getConcurrencyMode()).isTrue();
-
-    importStore.setConcurrencyMode(false);
-    assertThat(importStore.getConcurrencyMode()).isFalse();
-  }
-}
